created class HitFinder_beam (work in progress), changed OpWaveform to be a vecotr of doubles

+++ src/data/OpWaveform.cxx
@@ -3,8 +3,8 @@
 
 using namespace wcopreco;
 
-wcopreco::OpWaveform::OpWaveform(int ChanNum, double time_from_trig, int typ, const std::vector<short>& wfm)
- : std::vector<short> (wfm)
+wcopreco::OpWaveform::OpWaveform(int ChanNum, double time_from_trig, int typ, const std::vector<double>& wfm)
+ : std::vector<double> (wfm)
   {
     // OpWaveform = std::vector<short> empty_v(10,0);
     ChannelNum = ChanNum;
@@ -14,7 +14,7 @@ wcopreco::OpWaveform::OpWaveform(int ChanNum, double time_from_trig, int typ, co
 
 
 wcopreco::OpWaveform::OpWaveform(int ChanNum, double time_from_trig, int typ, const int nelems)
- : std::vector<short> (nelems,0)
+ : std::vector<double> (nelems,0)
   {
     // OpWaveform = std::vector<short> empty_v(10,0);
     ChannelNum = ChanNum;
